#8.14 Bugfix

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video, { formatHashtags } from "../models/Video";
     
 //global Router
@@ -17,24 +18,42 @@ export const search = async(req, res) => {
 //video Router
 export const watch = async(req, res) => {
     const {id} = req.params;
-    const video = await Video.findById({_id : id});
-    if(!video)
+    const video = await Video.findById(id).populate("owner");
+    if(!video){
         return res.status(400).render("error", {pageTitle : "Server Error"});
-    res.render("watch", {pageTitle : "watch", video});
+    }
+    return res.render("watch", {pageTitle : "watch", video});
 };
+
 export const getEdit = async(req, res) => {
+    const {
+        session : {
+            user : {_id},
+        }
+    } =req;
     const {id} = req.params;
-    const video = await Video.findById({_id : id});
+    const video = await Video.findById(id);
     if(!video)
         return res.status(400).render("error", {pageTitle : "Server Error"});
-    res.render("edit", {pageTitle : "Edit", video});
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
+    return res.render("edit", {pageTitle : "Edit", video});
 };
 
 export const postEdit = async(req, res) => {
+    const {
+        session :{
+            user : {_id},
+        }
+    }=req;
     const {id} = req.params;
-    const video = await Video.exists({_id : id});
+    const video = await Video.findById(id);
     if(!video)
         return res.status(400).render("error", {pageTitle : "Server Error"});
+    if(String(video.owner)!=String(_id)){
+        return res.status(403).redirect("/");
+    }
     const {title, description, hashtags} = req.body;
     await Video.findOneAndUpdate(id, {
         title,
@@ -49,15 +68,24 @@ export const getUpload = (req, res) =>{
 }
 
 export const postUpload = async(req, res)=>{
+    const {
+        session:{
+            user:{_id},
+        }
+    }=req;
     const {title, description, hashtags} = req.body;
     const {path} = req.file;
     try{
-        await Video.create({
+        const video = await Video.create({
+            owner : _id,
             fileUrl : path,
             title,
             description,
             hashtags : formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(video._id);
+        await user.save();
         res.redirect("/");
     }catch(error){
         console.log(error);
@@ -66,7 +94,19 @@ export const postUpload = async(req, res)=>{
 }   
 
 export const deleteVideo = async(req, res) => {
+    const {
+        session : {
+            user : {_id},
+        }
+    } =req;
     const {id} = req.params;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(400).render("error", {pageTitle : "Server Error"});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     res.redirect("/");
 }
